+++ pytest/test_hdbscan.py
@@ -26,7 +26,7 @@ path = "benchmark_data"
 def test_hdbscan():
     for dataset in ["jain", "pathbased"]:#, "s1", "Aggregation", "WUT_Smile", "unbalance", "a1"]:
         X = np.loadtxt("%s/%s.data.gz" % (path,dataset), ndmin=2)
-        labels = np.loadtxt("%s/%s.labels0.gz" % (path,dataset), dtype='int')
+        labels = np.loadtxt("%s/%s.labels0.gz" % (path,dataset), dtype=np.intc)
         label_counts = np.unique(labels,return_counts=True)[1]
         k = len(label_counts)
         D = scipy.spatial.distance.pdist(X)
